# smallville/src/main/java/io/github/nickm980/smallville/api/AgentStateResponse.java
package io.github.nickm980.smallville.api;

public class AgentStateResponse {
    private String name;
    private String action;
    private String location;
    private String emoji;
    
    public String getName() {
        return name;
    }
    public void setName(String name) {
        this.name = name;
    }
    public String getAction() {
        return action;
    }
    public void setAction(String action) {
        this.action = action;
    }
    public String getLocation() {
        return location;
    }
    
    public void setLocation(String location) {
        this.location = location;
    }
    
    public String getEmoji() {
        return emoji;
    }
    public void setEmoji(String emoji) {
        this.emoji = emoji;
    }
    
    
}