# smallville/src/main/java/io/github/nickm980/smallville/api/AskQuestionRequest.java
package io.github.nickm980.smallville.api;

public class AskQuestionRequest {

    private String question;

    public String getQuestion() {
	return question;
    }

    public void setQuestion(String question) {
	this.question = question;
    }

}