# smallville/src/main/java/io/github/nickm980/smallville/api/CreateLocationRequest.java
package io.github.nickm980.smallville.api;


public class CreateLocationRequest {

    private String name;

    public String getName() {
	return name;
    }

    public void setName(String name) {
	this.name = name;
    }
}