# smallville/src/main/java/io/github/nickm980/smallville/api/LocationStateResponse.java
package io.github.nickm980.smallville.api;

public class LocationStateResponse {

    private String name;
    private String state;

    public String getName() {
	return name;
    }

    public void setName(String name) {
	this.name = name;
    }

    public String getState() {
	return state;
    }

    public void setState(String state) {
	this.state = state;
    }
}